feat: refactor SaintBanner.vue
